Squashed commit of the following:
commit c892823492c0f094883e6160d25cf5c48d082ccd
Author: Victor Camargo <95081399+VictorSCamargo@users.noreply.github.com>
Date:   Tue Jan 17 11:43:32 2023 -0300

    Backend springboot reconstruido

commit fbb9219ad90a7155ad800992d316e9d5bab079aa
Author: Victor Camargo <95081399+VictorSCamargo@users.noreply.github.com>
Date:   Sun Jan 15 15:00:28 2023 -0300

    Desenvolvimento (#7)

    * Testando fetch para testagem

    * Register testes com mocked fetch criados

    Falta melhorar diversos detalhes. Tambem, criar arquivo com mensagens customizadas para padronizar as strings utilizadas no sistema

    * testes de componente Register completo

    incluindo simulacoes de fetch por meio de jest.spyOn

    * AlertMessages e SucessMessages

    Criada mensagens para manter um padrao nos elementos da aplicação. Falta adaptar o resto do sistema

    * Melhorias em Login e testes completos

    * Ajustes diversos e codigo de teste Alterar

    * Testes de unidade do componente Posts

    * Testes relacionados a rota e redirecionamento
