--- src/backend/postgresql-springboot/src/main/java/com/example/postgresqlspringboot/PostgresqlSpringbootApplication.java
@@ -1,13 +1,0 @@
-package com.example.postgresqlspringboot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PostgresqlSpringbootApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PostgresqlSpringbootApplication.class, args);
-	}
-
-}
