--- src/backend/postgresql-springboot/src/main/java/com/example/postgresqlspringboot/entity/ErrorMessage.java
@@ -1,21 +1,0 @@
-package com.example.postgresqlspringboot.entity;
-
-public class ErrorMessage {
-
-    private String message;
-
-    public ErrorMessage() {
-    }
-
-    public ErrorMessage(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
