--- src/backend/postgresql-springboot/src/main/java/com/example/postgresqlspringboot/entity/UserEntity.java
@@ -1,47 +1,0 @@
-package com.example.postgresqlspringboot.entity;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "users")
-public class UserEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String username;
-    private String password;
-
-    public UserEntity() {
-    }
-
-    public UserEntity(Long id, String username, String password) {
-        this.id = id;
-        this.username = username;
-        this.password = password;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
