--- src/backend/postgresql-springboot/src/main/java/com/example/postgresqlspringboot/entity/dto/UserDTO.java
@@ -1,31 +1,0 @@
-package com.example.postgresqlspringboot.entity.dto;
-
-public class UserDTO {
-
-    private String username;
-    private String password;
-
-    public UserDTO() {
-    }
-
-    public UserDTO(String username, String password) {
-        this.username = username;
-        this.password = password;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
