--- src/backend/postgresql-springboot/src/main/java/com/example/postgresqlspringboot/repository/PostRepository.java
@@ -1,9 +1,0 @@
-package com.example.postgresqlspringboot.repository;
-
-import com.example.postgresqlspringboot.entity.PostEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends JpaRepository<PostEntity, Long> {
-}
