--- src/backend/postgresql-springboot/src/main/java/com/example/postgresqlspringboot/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.example.postgresqlspringboot.repository;
-
-import com.example.postgresqlspringboot.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
-    List<UserEntity> findAllByUsername(String username);
-}
