--- src/backend/postgresql-springboot/src/main/resources/application.properties
@@ -1,12 +1,0 @@
-spring.jpa.database=POSTGRESQL
-spring.datasource.platform=postgres
-spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
-
-spring.datasource.username=postgres
-spring.datasource.password=bridge
-
-spring.jpa.show-sql=true
-spring.jpa.generate-ddl=true
-spring.jpa.hibernate.ddl-auto=create
-spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
-
