--- src/backend/postgresql-springboot/src/test/java/com/example/postgresqlspringboot/PostgresqlSpringbootApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.postgresqlspringboot;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PostgresqlSpringbootApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
