+++ src/backend/springboot-postgresql/src/main/java/com/example/demo/entity/ErrorMessage.java
@@ -1,0 +1,21 @@
+package com.example.demo.entity;
+
+public class ErrorMessage {
+
+    private String message;
+
+    public ErrorMessage() {
+    }
+
+    public ErrorMessage(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
