+++ src/backend/springboot-postgresql/src/main/java/com/example/demo/repository/PostRepository.java
@@ -1,0 +1,9 @@
+package com.example.demo.repository;
+
+import com.example.demo.entity.PostEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostRepository extends JpaRepository<PostEntity, Long> {
+}
