+++ src/backend/springboot-postgresql/src/main/java/com/example/demo/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.example.demo.repository;
+
+import com.example.demo.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+
+    List<UserEntity> findAllByUsername(String username);
+}
