+++ src/backend/springboot-postgresql/src/main/resources/application.properties
@@ -1,0 +1,11 @@
+spring.jpa.database=POSTGRESQL
+spring.datasource.platform=postgres
+spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
+
+spring.datasource.username=postgres
+spring.datasource.password=12345
+
+spring.jpa.show-sql=true
+spring.jpa.generate-ddl=true
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
